% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animate_bs.R
\name{animate_bs}
\alias{animate_bs}
\title{Make an animation of the model output in relation to bare soil.}
\usage{
animate_bs()
}
\value{
a GIF animated file
}
\description{
It requires ImageMagick installed in the system to work.
}
\examples{
## Not run animate_bs()
}
\keyword{animation}

