% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animate_model.R
\name{animate_model}
\alias{animate_model}
\title{Make an animation of the model output containing all state variables.}
\usage{
animate_model()
}
\value{
a GIF animated file
}
\description{
It requires ImageMagick installed in the system to work.
}
\examples{
## Not run animate_model()
}
\keyword{animation}

