% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animate_rb.R
\name{animate_rb}
\alias{animate_rb}
\title{Make an animation of the model output in relation to reed beds.}
\usage{
animate_rb()
}
\value{
a GIF animated file
}
\description{
It requires ImageMagick installed in the system to work.
}
\examples{
## Not run animate_rb()
}
\keyword{animation}

