% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animate_ss.R
\name{animate_ss}
\alias{animate_ss}
\title{Make an animation of the model output in relation to the salt steppe community.}
\usage{
animate_ss()
}
\value{
a GIF animated file
}
\description{
It requires ImageMagick installed in the system to work.
}
\examples{
## Not run animate_ss()
}
\keyword{animation}

