% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the model.}
\usage{
run_model(pgr_rb = 0.005, pgr_sm = 0.2, rnd = FALSE, method = "euler",
  TS = 0.25)
}
\arguments{
\item{pgr_rb}{potential growth rate of reed beds}

\item{pgr_sm}{potential growth rate of salt marsh}

\item{rnd}{create random initial state variables map. It is calculated based on a script adapted from Murray Efford (University of Auckland, New Zealand) and Santiago Saura (Universidad Politecnica de Madrid, Spain).}

\item{method}{integration method: "lsodes", "euler", "rk4", "ode23", "ode45", "adams", "iteration"}

\item{TS}{time step}
}
\value{
the function outputs a matrix named "out" which contains the model simulated values for every pixel, time step and state variable.
}
\description{
Model solving function using ode.2D from the 'deSolve' package.
}
\examples{
## Not run run_model()
}
\keyword{keywords}

